pyproject.toml
[tool.poetry]
name = "bankapp"
version = "0.1.0"
description = ""
authors = ["samolin <samolin.vlad@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.6"
django-rest-framework = "^0.1.0"
Pillow = "^9.2.0"
djoser = "^2.1.0"
django-filter = "^22.1"
django-bootstrap-datepicker-plus = "^4.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
